: Dozer Custom Converter
Task-Url:

+++ services/src/main/java/org/proto1/repository/CurrencyRepository.java
@@ -3,5 +3,5 @@ package org.proto1.repository;
 import org.proto1.domain.Currency;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CurrencyRepository extends CrudRepository<Currency, Long> {
+public interface CurrencyRepository extends CrudRepository<Currency, Integer> {
 }
